Refactor POI context analysis into dedicated services
